% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_partitions}
\alias{all_partitions}
\title{Build all sequential partitions.}
\usage{
all_partitions(n, kmax = n)
}
\arguments{
\item{n}{integer, sequence lenght to choose from.}

\item{kmax}{int, maximum number of segments in solution.}
}
\value{
list of all partitions.
}
\description{
Build all sequential partitions.
}
\examples{

all_partitions(4, 2)

}
