% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{const_cost}
\alias{const_cost}
\title{const_cost}
\usage{
const_cost(y, w, i, j)
}
\arguments{
\item{y}{NumericVector, values to group in order.}

\item{w}{NumericVector, weights.}

\item{i}{integer, first index (inclusive).}

\item{j}{integer, j>=i last index (inclusive);}
}
\value{
const cost of [i,...,j] interval (inclusive).
}
\description{
Calculate cost of using mean of points to estimate other points in interval.
Zero indexed.
}
\keyword{internal}
