% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{const_cost_logistic}
\alias{const_cost_logistic}
\title{const_cost_logistic}
\usage{
const_cost_logistic(y, w, min_seg, i, j)
}
\arguments{
\item{y}{NumericVector, 0/1 values to group in order.}

\item{w}{NumericVector, weights.}

\item{min_seg}{positive integer, minimum segment size.}

\item{i}{integer, first index (inclusive).}

\item{j}{integer, j>=i last index (inclusive);}
}
\value{
scalar, const cost of [i,...,j] interval (inclusive).
}
\description{
Calculate logistic cost of using mean of points to estimate other points in interval.
Zero indexed.
}
\examples{

const_cost_logistic(c(1, 1, 2, 2), c(1, 1, 1, 1), 1, 0, 3)

}
\keyword{internal}
