% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lin_cost}
\alias{lin_cost}
\title{lin_cost}
\usage{
lin_cost(x, y, w, i, j)
}
\arguments{
\item{x}{NumericVector, x-coords of values to group.}

\item{y}{NumericVector, values to group in order.}

\item{w}{NumericVector, weights.}

\item{i}{integer, first index (inclusive).}

\item{j}{integer, j>=i last index (inclusive);}
}
\value{
linear cost of [i,...,j] interval (inclusive).
}
\description{
Calculate cost of using linear model fit on points to estimate other points in the interval.
Zero indexed.
}
\keyword{internal}
