% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lin_cost_logistic}
\alias{lin_cost_logistic}
\title{lin_cost_logistic logistic deviance pricing}
\usage{
lin_cost_logistic(x, y, w, min_seg, i, j)
}
\arguments{
\item{x}{NumericVector, x-coords of values to group.}

\item{y}{NumericVector, values to group in order (should be in interval [0,1]).}

\item{w}{NumericVector, weights (positive).}

\item{min_seg}{positive integer, minimum segment size.}

\item{i}{integer, first index (inclusive).}

\item{j}{integer, j>=i last index (inclusive);}
}
\value{
scalar, linear cost of [i,...,j] interval (inclusive).
}
\description{
Calculate deviance cost of using linear model fit on points to estimate other points in the interval.
Zero indexed.
}
\details{
Note: this is the deviance cost of a linear fit, not the deviance loss of a logistic fit.
}
\examples{

lin_cost_logistic(c(1, 2, 3, 4), c(0.1, 0.2, 0.2, 0.1), c(1, 1, 1, 1), 1, 0, 3)

}
\keyword{internal}
