% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lin_costs}
\alias{lin_costs}
\title{lin_costs}
\usage{
lin_costs(x, y, w, min_seg, indices)
}
\arguments{
\item{x}{NumericVector, x-coords of values to group.}

\item{y}{NumericVector, values to group in order.}

\item{w}{NumericVector, weights.}

\item{min_seg}{positive integer, minimum segment size (>=1).}

\item{indices}{IntegerVector, ordered list of indices to pair.}
}
\value{
xcosts NumericMatix, for j>=i xcosts(i,j) is the cost of partition element [i,...,j] (inclusive).
}
\description{
Built matrix of interval costs for held-out linear models.
One indexed.
}
\examples{

lin_costs(c(1, 2, 3, 4), c(1, 2, 2, 1), c(1, 1, 1, 1), 1, 1:4)

}
