% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lin_costs_logistic}
\alias{lin_costs_logistic}
\title{lin_costs_logistic deviance costs.}
\usage{
lin_costs_logistic(x, y, w, min_seg, indices)
}
\arguments{
\item{x}{NumericVector, x-coords of values to group.}

\item{y}{NumericVector, values to group in order (should be in interval [0,1]).}

\item{w}{NumericVector, weights (should be positive).}

\item{min_seg}{positive integer, minimum segment size (>=1).}

\item{indices}{IntegerVector, ordered list of indices to pair.}
}
\value{
xcosts NumericMatix, for j>=i xcosts(i,j) is the cost of partition element [i,...,j] (inclusive).
}
\description{
Built matrix of interval deviance costs for held-out logistic models.
Fits are evaluated in-sample.
One indexed.
}
\examples{

lin_costs_logistic(c(1, 2, 3, 4, 5, 6, 7), c(0, 0, 1, 0, 1, 1, 0), c(1, 1, 1, 1, 1, 1, 1), 3, 1:7)

}
