% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtreat_coders.R
\name{piecewise_linear}
\alias{piecewise_linear}
\title{Piecewise linear fit.}
\usage{
piecewise_linear(varName, x, y, w = NULL)
}
\arguments{
\item{varName}{character, name of variable to work on.}

\item{x}{numeric, input values.}

\item{y}{numeric, values to estimate.}

\item{w}{numeric, weights.}
}
\description{
\code{vtreat} custom coder based on \code{RcppDynProg::solve_for_partition()}.
}
\examples{

piecewise_linear("x", 1:8, c(1, 2, 3, 4, 4, 3, 2, 1))

}
