% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RcppDynProg}
\alias{RcppDynProg}
\title{RcppDynProg}
\description{
Rcpp dynamic programming solutions for partitioning and machine learning problems.
Includes out of sample fitting applications.  
Also supplies additional custom coders for the vtreat package.
Please see \url{https://github.com/WinVector/RcppDynProg} for details.
}
\author{
John Mount
}
