% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{score_solution}
\alias{score_solution}
\title{compute the price of a partition solution (and check is valid).}
\usage{
score_solution(x, solution)
}
\arguments{
\item{x}{NumericMatix, for j>=i x(i,j) is the cost of partition element [i,...,j] (inclusive).}

\item{solution}{vector of indices}
}
\value{
price
}
\description{
compute the price of a partition solution (and check is valid).
}
\examples{

x <- matrix(c(1,1,5,1,1,0,5,0,1), nrow=3)
s <- c(1, 2, 4)
score_solution(x, s)

}
