% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_dynamic_program}
\alias{solve_dynamic_program}
\title{solve_dynamic_program}
\usage{
solve_dynamic_program(x, kmax)
}
\arguments{
\item{x}{NumericMatix, for j>=i x(i,j) is the cost of partition element [i,...,j] (inclusive).}

\item{kmax}{int, maximum number of segments in solution.}
}
\value{
dynamic program solution.
}
\description{
Solve a for a minimal cost partition of the integers [1,...,nrow(x)] problem where for j>=i x(i,j).
is the cost of choosing the partition element [i,...,j]. 
Returned solution is an ordered vector v of length k where: v[1]==1, v[k]==nrow(x)+1, and the 
partition is of the form [v[i], v[i+1]) (intervals open on the right).
}
