% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_interval_partition_no_k}
\alias{solve_interval_partition_no_k}
\title{solve_interval_partition interval partition problem, no boun on the number of steps.}
\usage{
solve_interval_partition_no_k(x)
}
\arguments{
\item{x}{NumericMatix, for j>=i x(i,j) is the cost of partition element [i,...,j] (inclusive).}
}
\value{
dynamic program solution.
}
\description{
Not working yet.
}
\details{
Solve a for a minimal cost partition of the integers [1,...,nrow(x)] problem where for j>=i x(i,j).
is the cost of choosing the partition element [i,...,j]. 
Returned solution is an ordered vector v of length k where: v[1]==1, v[k]==nrow(x)+1, and the 
partition is of the form [v[i], v[i+1]) (intervals open on the right).
}
\examples{

costs <- matrix(c(1.5, NA ,NA ,1 ,0 , NA, 5, -1, 1), nrow = 3)
solve_interval_partition(costs, nrow(costs))

}
