% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xlin_fits}
\alias{xlin_fits}
\title{xlin_fits}
\usage{
xlin_fits(x, y, w, i, j)
}
\arguments{
\item{x}{NumericVector, x-coords of values to group.}

\item{y}{NumericVector, values to group in order.}

\item{w}{NumericVector, weights.}

\item{i}{integer, first index (inclusive).}

\item{j}{integer, j>=i+2 last index (inclusive);}
}
\value{
vector of predictions.
}
\description{
Calculate out of sample linear fit predictions.
Zero indexed.
}
\keyword{internal}
