% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlin_fits_lm.R
\name{xlin_fits_lm}
\alias{xlin_fits_lm}
\title{xlin_fits_R}
\usage{
xlin_fits_lm(x, y, w)
}
\arguments{
\item{x}{NumericVector, x-coords of values to group (length>=2).}

\item{y}{NumericVector, values to group in order.}

\item{w}{NumericVector, weights (positive).}
}
\value{
vector of predictions.
}
\description{
Calculate out of sample linear fit predictions.
}
\examples{

xlin_fits_lm(c(1, 2, 3, 4), c(1, 2, 2, 1), c(1, 1, 1, 1))

}
\keyword{internal}
