% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xlin_pfits}
\alias{xlin_pfits}
\title{xlin_pfits}
\usage{
xlin_pfits(x, y, w, i, j)
}
\arguments{
\item{x}{NumericVector, explanatory variable (length>=2).}

\item{y}{NumericVector, values to fit.}

\item{w}{NumericVector, weights (positive).}

\item{i}{integer, first index (inclusive).}

\item{j}{integer, j>=i+2 last index (inclusive);}
}
\value{
vector of predictions.
}
\description{
Calculate out of sample linear fit predictions using pseudo-inverse.
Please see: \url{https://win-vector.com/2019/01/08/a-beautiful-2-by-2-matrix-identity/}.
Zero indexed.
}
\examples{

xlin_pfits(c(1, 2, 3, 4), c(1, 2, 2, 1), c(1, 1, 1, 1), 0, 3)

}
\keyword{internal}
