% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xlog_fits}
\alias{xlog_fits}
\title{Out of sample logistic predictions (in link space).}
\usage{
xlog_fits(x, y, w, i, j)
}
\arguments{
\item{x}{NumericVector, expanatory variable.}

\item{y}{NumericVector, 0/1 values to fit.}

\item{w}{NumericVector, weights (required, positive).}

\item{i}{integer, first index (inclusive).}

\item{j}{integer, last index (inclusive).}
}
\value{
vector of predictions for interval.
}
\description{
1-hold out logistic regression predections.
Zero indexed.
}
\examples{

set.seed(5)
d <- data.frame(
  x =  rnorm(10),
  y = sample(c(0,1), 10, replace = TRUE)
)
weights <- runif(nrow(d))
m <- glm(y~x, data = d, family = binomial, weights = weights)
coef(m)
xlog_fits(d$x, d$y, weights, 0, nrow(d)-1)

}
\keyword{internal}
