% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lin_cost}
\alias{lin_cost}
\title{lin_cost}
\usage{
lin_cost(x, y, w, min_seg, i, j)
}
\arguments{
\item{x}{NumericVector, x-coords of values to group.}

\item{y}{NumericVector, values to group in order.}

\item{w}{NumericVector, weights.}

\item{min_seg}{positive integer, minimum segment size.}

\item{i}{integer, first index (inclusive).}

\item{j}{integer, j>=i last index (inclusive);}
}
\value{
scalar, linear cost of [i,...,j] interval (inclusive).
}
\description{
Calculate cost of using linear model fit on points to estimate other points in the interval.
Zero indexed.
}
\examples{

lin_cost(c(1, 2, 3, 4), c(1, 2, 2, 1), c(1, 1, 1, 1), 1, 0, 3)

}
\keyword{internal}
